requirements.txt
numpy
pandas
streamlit
openai
python-dotenv
watchdog
pyodbc